Expose currentIndex as public.
Useful in case we want the underlying app to make decisions based on the tutorial manager's progression state (eg. especially if the tutorial was interrupted half-way through, causing the app to be quit, and we need to reconfigure the app's UI to whatever state it should have been at that tutorial stage.